pyproject.toml
[project]
name = "tradable-mind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-google-genai",
    "langchain",
    "langgraph",
    "twelvedata",
    "pandas",
    "python-dotenv",
    "matplotlib",
    "mplfinance"
]
